requirements.txt
streamlit==1.31.0
faiss-cpu==1.7.4
groq==0.4.0
numpy==1.24.3
python-dotenv==1.0.0